requirements.txt
transformers>=4.52.2
torch==2.6.0
torchvision==0.21.0
ray[default]==2.10.0
hydra-core>=1.3.0
omegaconf>=2.3.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
wandb>=0.15.0
rich>=13.0.0
hdbscan>=0.8.29
pandas>=2.0.0
tqdm>=4.65.0
pillow>=10.0.0
seaborn>=0.12.0
python-dotenv==1.1.0
tinyimagenet 